task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,4 @@
 #include "function_pointers.h"
-
 /**
  * array_iterator - executes a function given as a parameter
  * @array: pointer to an array
@@ -8,7 +7,7 @@
  * Return: void
  */
 
-void array_iterator(int *array, size_t size, void(*action)(int))
+void array_iterator(int *array, int size, void(*action)(int))
 {
 	int i = 0;
